Rename file to match folder
